Fixed bug where you couldn't succesfully POST when database table was empty. + Made constraint so you cant delete locations or categories when in use by an Item (todo: notification to user when this happens)

+++ InventoryManagement.Api/Controllers/CategoryController.cs
@@ -93,6 +93,13 @@ namespace InventoryManagement.Api.Controllers
         {
             try
             {
+                var items = await this.itemRepository.GetItems();
+                var categoryIsUsed = items.Any(x => x.CategoryId == id);
+                if (categoryIsUsed)
+                {
+                    return Forbid();
+                }
+
                 var category = await this.itemRepository.DeleteCategory(id);
 
                 if (category == null)
